+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Installer/EnvironmentInstaller.cs
@@ -8,10 +8,17 @@ namespace Baku.VMagicMirror.Installer
     [RequireComponent(typeof(PostProcessLayer))]
     public class EnvironmentInstaller : InstallerBase
     {
+        [SerializeField] private Camera mainCamera;
+        [SerializeField] private Camera refCameraForRay;
         public override void Install(DiContainer container)
         {
-            container.Bind<Camera>().FromMethod(_ => GetComponent<Camera>()).AsCached();
+            container.BindInstance(mainCamera);
+            container.BindInstance(refCameraForRay).WithId("RefCameraForRay");
             container.Bind<PostProcessLayer>().FromMethod(_ => GetComponent<PostProcessLayer>()).AsCached();
+            container.BindInterfacesTo<CameraFovController>().AsSingle();
+            container.Bind<CameraUtilWrapper>().AsSingle();
+
+            container.BindInterfacesTo<AntiAliasSettingSetter>().AsSingle();
         }
     }
 }