+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/Interprocess/Model/VmmCommands.cs
@@ -150,6 +150,7 @@
 
         // Image Quality
         public const string SetImageQuality = nameof(SetImageQuality);
+        public const string SetAntiAliasStyle = nameof(SetAntiAliasStyle);
         public const string SetHalfFpsMode = nameof(SetHalfFpsMode);
 
         // Lighting 
@@ -214,7 +215,11 @@
         public const string RequestResetAllAccessoryLayout = nameof(RequestResetAllAccessoryLayout);
         public const string RequestResetAccessoryLayout = nameof(RequestResetAccessoryLayout);
         public const string ReloadAccessoryFiles = nameof(ReloadAccessoryFiles);
-            
+        
+        // External Video Sharing
+        public const string EnableSpoutOutput = nameof(EnableSpoutOutput);
+        public const string SetSpoutOutputResolution = nameof(SetSpoutOutputResolution);
+        
         // VRoidHub
         public const string OpenVRoidSdkUi = nameof(OpenVRoidSdkUi);
         public const string RequestLoadVRoidWithId = nameof(RequestLoadVRoidWithId);