+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/TextureSharing/SpoutSenderInstaller.cs
@@ -1,0 +1,16 @@
+using UnityEngine;
+using Zenject;
+
+namespace Baku.VMagicMirror
+{
+    public class SpoutSenderInstaller : MonoInstaller
+    {
+        [SerializeField] private SpoutSenderWrapperView spoutSenderView;
+
+        public override void InstallBindings()
+        {
+            Container.BindInstance(spoutSenderView);
+            Container.BindInterfacesAndSelfTo<SpoutSenderController>().AsSingle();
+        }
+    }
+}